bug修复

+++ SVT/Generation/Entry.h
@@ -12,7 +12,7 @@ namespace GenMain
 
 	extern std::map<std::pair<int, int>, Generation::WorldUnit*> WorldUnitTmp;
 
-	extern unsigned short* WorldBlock(int x, int y, int z);
+	extern unsigned short* WorldBlock(float x, float y, float z);
 
 	extern inline void RegisterBiomeMain();
