+++ SVT/UCLient/Graph/BasicClass/lib/BasicHeaders.h
@@ -21,5 +21,5 @@ extern void glClearError();
 extern void glCheckError();
 extern bool glLogCall(const char* function, const char* file, int line);
 
-extern int IntDiv(int a, int b);
-extern int IntMod(int a, int b);
+extern int IntDiv(float a, int b);
+extern int IntMod(float a, int b);
